Fix build errors and verify existing features
- Fix Go type mismatches in category_service.go, user_group_service.go,
  subscription_service.go, book_access_service.go, reading_session_service.go
- Fix TypeScript errors: API functions return data directly, not wrapped
- Update Layout component to accept children prop
- All features already implemented:
  * Gmail-only registration validation
  * User groups (students, free, subscribers)
  * Book file upload/storage (PDF, EPUB, MOBI)
  * Time-limited access (14 days default)
  * Online PDF reader with progress tracking
  * Reading sessions and statistics

+++ internal/services/book_access_service.go
@@ -127,7 +127,7 @@ func (s *bookAccessService) UpdateProgress(id uuid.UUID, currentPage int, readTi
 	}
 
 	access.CurrentPage = currentPage
-	access.TotalReadTime += int64(readTime.Seconds())
+	access.TotalReadTime += int(readTime.Seconds())
 	now := time.Now()
 	access.LastAccessedAt = &now
